pyproject.toml
[project]
name = "daily-intelligence-newsletter"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.32",
  "pyyaml>=6.0",
  "httpx>=0.27",
  "feedparser>=6.0",
  "apscheduler>=3.10",
  "pydantic>=2.0",
  "pydantic-settings>=2.0",
  "google-genai>=1.0",
  "openai>=1.0",
  "python-dotenv>=1.0",
  "playwright>=1.49",
  "parsel>=1.9",
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["pipeline*", "tools*", "backend*"]
